+++ apps/laboratory/src/components/Wagmi/WagmiCreatePasskeySignerTest.tsx
@@ -4,10 +4,10 @@ import { privateKeyToAccount } from 'viem/accounts'
 import { useChakraToast } from '../Toast'
 import { LOCAL_SIGNER_KEY, getItem } from '../../utils/LocalStorage'
 import { createCredential } from 'webauthn-p256'
-import { useWagmiPermissionsSync } from '../../context/WagmiPermissionsSyncContext'
+import { usePasskey } from '../../context/PasskeyContext'
 
 export function WagmiCreatePasskeySignerTest() {
-  const { isPasskeyAvailable, setPasskey, passkeyId } = useWagmiPermissionsSync()
+  const { isPasskeyAvailable, setPasskey, passkeyId } = usePasskey()
   const toast = useChakraToast()
 
   async function handleCreatePasskey() {
