+++ apps/laboratory/src/components/Wagmi/WagmiRequestPermissionsAsyncTest.tsx
@@ -3,112 +3,23 @@ import { useAccount } from 'wagmi'
 import { walletActionsErc7715 } from 'viem/experimental'
 import { useCallback, useState } from 'react'
 import { useChakraToast } from '../Toast'
-import { createPublicClient, custom } from 'viem'
+import { createWalletClient, custom, type Chain } from 'viem'
 import { EIP_7715_RPC_METHODS } from '../../utils/EIP5792Utils'
-import { useWalletConnectCosigner } from '../../hooks/useWalletConnectCosigner'
-import { useWagmiAvailableCapabilities } from '../../hooks/useWagmiActiveCapabilities'
-import { useWagmiPermissionsAsync } from '../../context/WagmiPermissionsAsyncContext'
 import {
-  decodeUncompressedPublicKey,
-  encodePublicKeyToDID,
-  hexStringToBase64
-} from '../../utils/EncodingUtils'
+  useWagmiAvailableCapabilities,
+  type Provider
+} from '../../hooks/useWagmiActiveCapabilities'
+import { useLocalEcdsaKey } from '../../context/LocalEcdsaKeyContext'
 import { bigIntReplacer } from '../../utils/CommonUtils'
-import { getSampleAsyncPermissions } from '../../utils/ERC7715Utils'
+import { useERC7715Permissions } from '../../hooks/useERC7715Permissions'
 
 export function WagmiRequestPermissionsAsyncTest() {
   const { provider, supported } = useWagmiAvailableCapabilities({
     method: EIP_7715_RPC_METHODS.WALLET_GRANT_PERMISSIONS
   })
   const { chain, address, isConnected } = useAccount()
-  const caip10Address = `eip155:${chain?.id}:${address}`
 
-  const {
-    projectId,
-    signer,
-    grantedPermissions,
-    clearGrantedPermissions,
-    setGrantedPermissions,
-    setWCCosignerData
-  } = useWagmiPermissionsAsync()
-  const [isRequestPermissionLoading, setRequestPermissionLoading] = useState<boolean>(false)
-  const { addPermission, updatePermissionsContext } = useWalletConnectCosigner(
-    caip10Address,
-    projectId
-  )
-  const toast = useChakraToast()
-
-  const onRequestPermissions = useCallback(async () => {
-    setRequestPermissionLoading(true)
-
-    if (!signer) {
-      throw new Error('PrivateKey signer not available')
-    }
-    if (!provider) {
-      throw new Error('No Provider available, Please connect your wallet.')
-    }
-    try {
-      const addPermissionResponse = await addPermission({
-        permissionType: 'donut-purchase',
-        data: '',
-        onChainValidated: false,
-        required: true
-      })
-      setWCCosignerData(addPermissionResponse)
-      const cosignerPublicKey = decodeUncompressedPublicKey(addPermissionResponse.key)
-
-      const dAppECDSAPublicKey = signer.publicKey
-      const dAppSecp256k1DID = encodePublicKeyToDID(dAppECDSAPublicKey, 'secp256k1')
-      const coSignerSecp256k1DID = encodePublicKeyToDID(cosignerPublicKey, 'secp256k1')
-
-      const publicClient = createPublicClient({
-        chain,
-        // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-        transport: custom(provider)
-      }).extend(walletActionsErc7715())
-      const samplePermissions = getSampleAsyncPermissions([coSignerSecp256k1DID, dAppSecp256k1DID])
-      const approvedPermissions = await publicClient.grantPermissions(samplePermissions)
-      if (approvedPermissions) {
-        await updatePermissionsContext({
-          pci: addPermissionResponse.pci,
-          context: {
-            expiry: approvedPermissions.expiry,
-            signer: {
-              type: 'donut-purchase',
-              data: {
-                ids: [addPermissionResponse.key, hexStringToBase64(dAppECDSAPublicKey)]
-              }
-            },
-            signerData: {
-              // eslint-disable-next-line @typescript-eslint/no-non-null-assertion, @typescript-eslint/no-non-null-asserted-optional-chain
-              userOpBuilder: approvedPermissions.signerData?.userOpBuilder!
-            },
-            permissionsContext: approvedPermissions.permissionsContext,
-            factory: approvedPermissions.factory || '',
-            factoryData: approvedPermissions.factoryData || ''
-          }
-        })
-        setGrantedPermissions(approvedPermissions)
-        setRequestPermissionLoading(false)
-        toast({
-          type: 'success',
-          title: 'Permissions Granted',
-          description: JSON.stringify(approvedPermissions, bigIntReplacer)
-        })
-
-        return
-      }
-      toast({ title: 'Error', description: 'Failed to obtain permissions' })
-    } catch (error) {
-      toast({
-        type: 'error',
-        title: 'Permissions Erros',
-        description: error instanceof Error ? error.message : 'Some error occurred'
-      })
-    }
-    setRequestPermissionLoading(false)
-  }, [signer, provider])
-  if (!isConnected || !provider || !address) {
+  if (!isConnected || !provider || !address || !chain) {
     return (
       <Text fontSize="md" color="yellow">
         Wallet not connected
@@ -123,14 +34,56 @@ export function WagmiRequestPermissionsAsyncTest() {
     )
   }
 
+  return <ConnectedTestContent chain={chain} provider={provider} />
+}
+
+function ConnectedTestContent({ chain, provider }: { chain: Chain; provider: Provider }) {
+  const { grantedPermissions, clearGrantedPermissions, requestPermissionsAsync } =
+    useERC7715Permissions({ chain })
+  const { signer } = useLocalEcdsaKey()
+  const [isRequestPermissionLoading, setRequestPermissionLoading] = useState<boolean>(false)
+  const toast = useChakraToast()
+  const onRequestPermissions = useCallback(async () => {
+    setRequestPermissionLoading(true)
+
+    if (!signer) {
+      throw new Error('PrivateKey signer not available')
+    }
+    if (!provider) {
+      throw new Error('No Provider available, Please connect your wallet.')
+    }
+
+    const walletClient = createWalletClient({
+      chain,
+      transport: custom(provider)
+    }).extend(walletActionsErc7715())
+
+    const response = await requestPermissionsAsync(walletClient, signer)
+    if ('error' in response) {
+      toast({
+        type: 'error',
+        title: 'Request Permissions Errors',
+        description: response.message
+      })
+      setRequestPermissionLoading(false)
+
+      return
+    }
+    toast({
+      type: 'success',
+      title: 'Permissions Granted',
+      description: JSON.stringify(response.approvedPermissions, bigIntReplacer)
+    })
+
+    setRequestPermissionLoading(false)
+  }, [signer, provider])
+
   return (
     <Stack direction={['column', 'column', 'row']}>
       <Button
         data-test-id="request-permissions-button"
         onClick={onRequestPermissions}
-        isDisabled={Boolean(
-          isRequestPermissionLoading || Boolean(grantedPermissions) || !isConnected
-        )}
+        isDisabled={Boolean(isRequestPermissionLoading || Boolean(grantedPermissions))}
         isLoading={isRequestPermissionLoading}
       >
         Request Permissions
